Update package source name to "IG ProGet IgniteUINuGet - Staging"
Seeing an error authenticating to the staging feed. While it worked recently, it may be an issue with a naming mismatch between the package source created in the nuget.config file we generate and the name for the service connection used for credentials that didn't appear until after a system reboot.